Huge refactoring
Unfortunately, ORM-code becomes hard to extend, if it splitted into property-classes. Much easier (and probable more performance-friendly) to generate this code directly.
It becomes clear with caches - it looks ugly, if generated in properties, but it depends on type of ID field.
It is also looks easier to support (in case of a good test coverage).