Equals for null parameters
In SQLite NULL = NULL returns false, but NULL IS NULL - true. It is unexpected when you just pass null parameters to the query method. That is why by default all operators '=' and '==' will be replaced to IS and operators '<>' and '!=' to IS NOT. To prevent this behaviour just set fixNullableEquals to false.